Handle events that span months
The library requests events from the app in month chunks. It does this
for the current month, previous month, and next month, upon month
change.  However, events that span month boundaries will be returned
by the app in all callbacks for months in which the event exists. This
can result in the libaries's month-based cache having the same event
multiple times. We need make sure the event only appears in the
earliest month it applies to, or the library will render it twice.